# Significantly adjusted ASMD recipes