~~Balance adjusts them to be less silly. Costs down below in terms of ingots:~~
RECIPES HAVE BEEN CHANGED SINCE, SEE BELOW.
ASMD Resistors: 4 PBI, 1 Graphene, 2 Platinum
ASMD Diodes: 4 PBI, 1 Indium Gallium Phosphide, 1 Niobium Titanium
ASMD Transistors: 4 PBI, 1 Vanadium Gallium, 2 HSSG
ASMD Capacitors: 5 PBI (4 foil), 2 HSSE
All of them have 300 ticks or 15 seconds at EV and require 576L or 4 ingots of PBI. Each recipe gives 64 ASMDs.
The Wrapper recipe requires 16x the materials but gives 64 Wrappers. Additionally, they require LuV voltage at 300 ticks or 15 seconds.

~~These might be a tiny bit cheap? Please do give your thoughts, any will be greatly appreciated to find out the general idea of how much these should cost. I didn't want to go overboard or make them too obscene. Additionally, I tried to fix the inherent problem of them being too PBI hungry wise.~~