@POPlol333 2021-11-11
They aren't that cheap tbh, I'm about to enter LuV and I can't really afford them in decent quantity