@ScriptedPiky 2021-11-11
Oh yeah I should also clarify that they do give a stack of 64 ASMDs per recipe process. So in this case, it may or may not be cheap per amount of ASMDs you may potentially get.