@ScriptedPiky 2021-11-11
> I think, nerfing the quantity instead of the materials they ask would be a better idea, like 32 or 48 maybe?

Yeah, I think that would be best here. 32 maybe sounds good?