@GlodBlock 2021-11-12
could you add a buffed recipe with more advanced material, like using tantalum in SMD capacitor to get more capacitors.