@ScriptedPiky 2021-11-12
> could you add a buffed recipe with more advanced material, like using tantalum in SMD capacitor to get more capacitors.

Unsure how I would go about this, both implementing and material wise but that also could also be nice if I shifted the materials slightly, though ultimately unsure.

> Can we please get rid of the Control circuits?

That's on Dream's end. I don't want to remove it without his request, though I very much would remove it as it's very frustrating for any sort of crafting, especially auto-crafting, and doesn't conflict much of anything.