@ArtherCasan 2021-11-12
Can we use the Rhodium-Plated Palladium instead of blue steel. It’s an advanced material and also gives usage of Rhodium and Palladium.