@Skiftcha 2021-11-12
any rfc issue linked? or discord discussion?
lets just add random materials into recipes?
why are they like that?
capacitor was conductor foil and dielectric foil. like in real capacitors. now just random metal bolts and foil.