@GTNH-Colen 2021-11-12
> Can we use the Rhodium-Plated Palladium instead of blue steel. It’s an advanced material and also gives usage of Rhodium and Palladium.

Seconded on this, rhodium is desperately underused. I mean you use it for LuV hulls and that's basically it from what I can remember?