@Florexiz 2021-11-12
Why we even changing this? Right now this feels like a big nerf without no reason and sense. SMDs aren't supposed to be expensive