@ScriptedPiky 2021-11-12
> Can we use the Rhodium-Plated Palladium instead of blue steel. It’s an advanced material and also gives usage of Rhodium and Palladium.

Sure, actually yeah, I would love to add those. Not sure about implementing them since I don't know if it registers Bartworks materials inside? I'm genuinely curious.

> any rfc issue linked? or discord discussion? lets just add random materials into recipes? why are they like that? capacitor was conductor foil and dielectric foil. like in real capacitors. now just random metal bolts and foil.

https://github.com/GTNewHorizons/GT-New-Horizons-Modpack/issues/7573 Original ticket on it. Though also yeah, the materials were thrown in. Any ideas?

> Why we even changing this? Right now this feels like a big nerf without any reason and sense. SMDs aren't supposed to be expensive

The current suggested recipes aren't even that bad comparably to the original ASMDs. They're not super heavy time time gates nor are they that difficult (except iridium or naquadah post-rework or some others), and I don't understand where this "big nerf without any reason and sense" comes from at all. The original ASMD recipes (apparently they recently were updated to be a stack of 64 instead of their original values?) were FAR worse, tedious, and obnoxiously difficult. The current suggested recipes are far tamer, though in an incredibly rough draft state.