@Retster 2021-11-12
> Balance adjusts them to be less silly. Costs down below in terms of ingots: ASMD Resistor: 2 Naquadah, 2 HSSE, 4 Yttrium Barium Cuprate ASMD Diode: 2 Indium Gallium Phosphide, 2 Iridium, 2 Niobium Titanium, 2 Vivid Alloy ASMD Transistor: 2 Vibrant Alloy, 2 Vanadium Gallium, 2 HSSG ASMD Capacitor: 2 HSSE, 2 Enderium, 4 Blue Steel All of them have 300 ticks or 15 seconds at EV and require 576L or 4 ingots of PBI. Each recipe gives 32 ASMDs; originally was 64.
> 
> These might be a tiny bit cheap? Please do give your thoughts, any will be greatly appreciated to find out the general idea of how much these should cost. I didn't want to go overboard or make them too obscene. Additionally, I tried to fix the inherent problem of them being too PBI hungry wise.


Hey, guys what is going here? This is advanced tech pack, or kindergarten?
I am against this approach, when changes are proposed without calculations, justification and comparison of what was and what was. 

When I make a choice between two recipes in the game(smd or asmd), I always make calculations and look at the consumption of critical materials that are difficult to obtain. Such as indium, ostium, etc. Next, I make a comparison of the recipes for SMD and ASMD as a whole, since you cannot use both in one recipe. And if the consumption of critical materials is too high, then I fucked your ASMD.

Make calculations and justification for the changes you are proposing or close your request. Thanks.