@Sphyix 2021-11-12
I don't understand why everything keeps getting nerfed for no apparent reason.
How grindy do you want this game to become?
I keep seeing pull requests and issues about things being "too easy" only, mostly by players that I don't even know if they got to endgame yet.

What about adding some new content instead of making every tier take twice as much time for no reason other than making the whole process of getting late game be longer?

Apart from the rant on top, I don't know if you have an idea on how much impact a change like this has with the game in general, those ASMD are used everywhere in high quantities, it changes completely the balance of making cirucits.
So if you want a change like this, make a spreadsheet with calculations for everything that uses ASMD or SMD to be able to check the difference between them.

What is the objective here?
ASMDs should be cheaper than SMDs? Yes/No, WHY?

 


> Rhodium-palladium foil and you've got my vote.

OH YES, make me spam more platline, I'll call you when delta lags like crazy cause everyone keeps using mats from the platline everywhere without fixing the initial issue of the platline being shit in terms of materials output.