@Sphyix 2021-11-12
> > OH YES, make me spam more platline, I'll call you when delta lags like crazy cause everyone keeps using mats from the platline everywhere without fixing the initial issue of the platline being shit in terms of materials output.
> 
> If you have valid proof of this, I'd gladly love to know, cause I'd be damned if a line of LCRs and PAs are going to lag more than giant filled ME networks or an EiO tree farm. Also, if you apply this logic to the Naquadah Rework, then I'm sorry to say that you're going to have to cancel that PR as well.

With you, I meant colen if that wasn't clear, anyway a plat line on delta adds a few ms, and that's a basic platline at IV-LuV with LCRs, avoiding PAs. Overclock on the platline stops at IV-LuV so can't go any faster. That can't supply a UHV+ player.
To be able to make a decent plat line right now I need to spam PAs (what a surprise). Each PA is around 0.1 to 0.2ms when running with a stack of LuV-ZPM machines inside it (Yes, LuV or ZPM machines are needed, maybe UV, to reach a decent speed compared to the perfectly overclocking LCR, PAs don't perfectly overclock)
So a fast platline can end up costing 5 to 10ms to a server like delta.

A thing to add, you doubled the cost of indium on the diodes and you claim it to be less tedious.
https://github.com/GTNewHorizons/GT-New-Horizons-Modpack/issues/7904
Indium right now is something that can be obtained in high quantities only with UU before fusion T3. Making the consumption of it double on circuits is not a good idea.